refactor: 클릭 이벤트 핸들러로 분리, styled.tsx 분리, custom hook 분리
